Define download data from API in ApiConsumerService
