+++ web/src/main/java/com/infoshareacademy/gitloopersi/mapper/HolidayMapper.java
@@ -1,0 +1,5 @@
+package com.infoshareacademy.gitloopersi.mapper;
+
+public class HolidayMapper {
+
+}
